setting ssh tunnel in separate thread and other small corrections

+++ Driver/testdata_sender.py
@@ -2,6 +2,7 @@ import subprocess
 from multiprocessing.connection import Client
 import time
 import pandas as pd
+import threading
 
 USE_REMOTE = False
 
@@ -30,8 +31,8 @@ class LivePlotter_Sender:
         # Define the SSH command to establish the tunnel
         ssh_command = [
             'ssh',
-            '-L', f"{self.address[1]}:{self.address[0]}:{self.address[1]}",  # Local port 6000 to remote port 6000
-            f"{self.remote_username}@{self.remote_ip}"  # Replace with your username and remote host
+            '-L', f"{self.address[1]}:{self.address[0]}:{self.address[1]}",  # Local port to remote port
+            f"{self.remote_username}@{self.remote_ip}"  # Username and remote host
         ]
 
         # Start the SSH tunnel as a subprocess
@@ -39,21 +40,28 @@ class LivePlotter_Sender:
         return process
 
     def connect(self):
+        thread = threading.Thread(target=self._connect)
+        thread.start()
+
+    def _connect(self):
         if self.use_remote:
             self.ssh_process = self.establish_ssh_tunnel()
-            time.sleep(5)
+            time.sleep(5)  # Simulate waiting for SSH tunnel to establish
         else:
-            time.sleep(3)
+            time.sleep(3)  # Simulate waiting for local setup
         self.connection = Client(self.address)
         self.connection_ready = True
 
     def send(self, data):
-        pass
+        if not self.connection_ready:
+            raise Exception("Connection not established yet.")
+        self.connection.send(data)
 
     def close(self):
-        self.connection.close()  # Close the connection
+        if self.connection:
+            self.connection.close()  # Close the connection
 
-        if self.use_remote:
+        if self.use_remote and self.ssh_process:
             # Terminate the SSH tunnel
             self.ssh_process.terminate()
             self.ssh_process.wait()  # Wait for the SSH process to terminate
@@ -72,23 +80,21 @@ def main():
     path = './ExperimentRecordings/CPP_mpc__2024-07-01_00-52-52.csv'
     # path = '../../../ExperimentRecordings/CPP_mpc__2024-07-01_00-52-52.csv'
     df = pd.read_csv(path, comment='#')
-
-
     df = df[['angle', 'angleD', 'position', 'positionD']]
 
     try:
         # Send the header
         header = df.columns.tolist()
-        sender.connection.send(header)
+        sender.send(header)
         time.sleep(0.1)
 
         # Send data line by line
         for _, row in df.iterrows():
-            sender.connection.send(row.values)
+            sender.send(row.values)
             time.sleep(0.1)  # Wait for a short time before sending the next row
 
         # Optionally, send a 'complete' message to indicate end of data
-        sender.connection.send('complete')
+        sender.send('complete')
 
     finally:
         sender.close()
